+++ pop-api/src/v0/fungibles/traits.rs
@@ -38,6 +38,7 @@ pub trait Psp22 {
 	/// - `from` - The account from which the token balance will be withdrawn.
 	/// - `to` - The recipient account.
 	/// - `value` - The number of tokens to transfer.
+	/// - `data` - Additional data with unspecified format.
 	#[ink(message)]
 	fn transfer_from(
 		&mut self,
@@ -72,57 +73,6 @@ pub trait Psp22 {
 	/// - `value` - The number of tokens to decrease the allowance by.
 	#[ink(message)]
 	fn decrease_allowance(&mut self, spender: AccountId, value: Balance) -> PSP22Result<()>;
-}
-
-/// The PSP22 Metadata trait.
-#[ink::trait_definition]
-pub trait Management {
-	/// Creates a new token with the specified `id`, assigns `admin` as the administrator,
-	/// and sets a minimum balance requirement.
-	///
-	/// # Parameters
-	/// - `id`: The unique identifier for the token to be created.
-	/// - `admin`: The account that will have administrative privileges over the token.
-	/// - `min_balance`: The minimum balance required to hold the token.
-	#[ink(message)]
-	fn create(&mut self, id: TokenId, admin: AccountId, min_balance: Balance) -> PSP22Result<()>;
-
-	/// Initiates the process to destroy the specified `token`.
-	///
-	/// # Parameters
-	/// - `token`: The identifier of the token to be destroyed.
-	#[ink(message)]
-	fn start_destroy(&mut self, token: TokenId) -> PSP22Result<()>;
-
-	/// Sets metadata for the specified `token`, including its name, symbol, and decimal places.
-	///
-	/// # Parameters
-	/// - `token`: The identifier of the token to set metadata for.
-	/// - `name`: The name of the token.
-	/// - `symbol`: The symbol representing the token.
-	/// - `decimals`: The number of decimal places the token uses.
-	#[ink(message)]
-	fn set_metadata(
-		&mut self,
-		token: TokenId,
-		name: Vec<u8>,
-		symbol: Vec<u8>,
-		decimals: u8,
-	) -> PSP22Result<()>;
-
-	/// Clears the metadata for the specified `token`.
-	///
-	/// # Parameters
-	/// - `token`: The identifier of the token to clear metadata for.
-	#[ink(message)]
-	fn clear_metadata(&mut self, token: TokenId) -> PSP22Result<()>;
-
-	/// Checks whether a token with the specified `token` identifier exists.
-	///
-	/// # Parameters
-	/// - `token`: The identifier of the token to check for existence.
-	#[ink(message)]
-	fn token_exists(&self, token: TokenId) -> PSP22Result<bool>;
 }
 
 /// The PSP22 Metadata trait.
